backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# YAML processing
pyyaml==6.0.1

# Google Gemini LLM
google-generativeai==0.3.2

# Environment management
python-dotenv==1.0.0

# HTTP client for API calls
httpx==0.26.0

# Date/time utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# CORS middleware
starlette==0.35.1
